feat: add raw material management with CRUD operations and update product service for price handling

+++ AutoflexBack/src/services/product.service.ts
@@ -1,3 +1,4 @@
+import { AppError } from '../errors/AppError';
 import { ProductRepository } from '../repositories/product.reposiory';
 
 interface CreateProductDTO {
@@ -10,17 +11,34 @@ const repository = new ProductRepository();
 
 export class ProductService {
   async create(data: CreateProductDTO) {
-    return repository.create(data);
+    const product = await repository.create(data);
+    const parsedProduct = { ...product, price: Number(product.price) }
+    return parsedProduct;
   }
 
   async list() {
-    return repository.list();
+    const products = await repository.list();
+    const parsedProducts = products.map(product => ({
+      ...product,
+      price: Number(product.price)
+    }));
+    return parsedProducts;
   }
 
   async deleteByCode(code: string) {
+    const product = await this.findByCode(code);
+    if (!product) {
+      throw new AppError('Product not found', 404);
+    }
     await repository.delete(code);
   }
-  async findByCode(code: string){
-    return repository.findByCode(code);
+
+  async findByCode(code: string) {
+    const product = await repository.findByCode(code);
+    if (!product) {
+      throw new AppError('Product not found', 404);
+    }
+    const parsedProduct = { ...product, price: Number(product?.price) };
+    return parsedProduct;
   }
 }
